requirements.txt
numpy==1.24.0
Pillow==9.4.0
scipy==1.10.1
streamlit==1.16.0
